asm_avr.inc: add missing memory clobber
It seems that older versions of GCC did not optimize memory accesses
across function calls. That made function calls an implicit memory
barrier, even when the function implementation was part of the current
compilation unit and the compiler could infer that a function did not
touch memory.

With inline assembly, the compiler assumes the clobber list to be
exhaustive. If no memory clobber is given and a function does not touch
memory except for the inline assembly, the compiler may assume that
a function does not change memory contents and optimize accordingly.

Adding the required memory clobbers fixes incorrect computations of
`vli_mmod_fast_secp256r1()` on AVR GCC 15.2.0.